LayoutTests/fast/canvas/svg-taint-expected.txt

CONSOLE MESSAGE: Unable to get image data from canvas because the canvas has been tainted by cross-origin data.
Let's check that rendering an SVG pattern to a canvas taints it!
See https://bugs.webkit.org/show_bug.cgi?id=36838

Starting...
Exception: SECURITY_ERR
 
